python/requirements.txt
python-dotenv
websocket-client==0.44.0